Followup work.
